Add files via upload
The SoCal RUG Hackathon 2022-04, hosted by the Southern California R Users Group and the UCI Paul Merage School of Business. It is a two-day event where you will "hack" a data set for fun, education, and prizes. The focus of the event is on education and teamwork, with the main goal of taking a data set from its "raw" form all the way through to a final "product" (e.g. visualization, model, insight).